backend/requirements.txt
asgiref==3.8.1
dj-rest-auth==6.0.0
Django==4.2.13
django-cors-headers==4.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
numpy==1.26.4
PyJWT==2.8.0
scipy==1.13.1
sqlparse==0.5.0
typing_extensions==4.12.0
